Moved menubar to its own component
