+++ src/features/appBar/MyAppBar.tsx
@@ -8,11 +8,18 @@ import DarkThemeToggler from '../darkThemeProvider/DarkThemeToggler';
 import { Location, useLocation } from 'react-router';
 import MyDrawer from './MyDrawer';
 import LoginButton from '../auth/LoginButton';
+import MenuBar from '../menuBar/MenuBar';
+import { useSelector } from 'react-redux';
+import { RootState } from '../../store';
 
 type Props = {};
 
 const MyAppBar = (props: Props) => {
     const location = useLocation();
+    const authed = useSelector(
+        (state: RootState) =>
+            state.authReducer.token != null && !state.authReducer.loading
+    );
 
     return (
         <Box sx={{ flexGrow: 1 }}>
@@ -27,7 +34,7 @@ const MyAppBar = (props: Props) => {
                         {translateLocationPathname(location)}
                     </Typography>
                     <DarkThemeToggler />
-                    <LoginButton />
+                    {authed ? <MenuBar /> : <LoginButton />}
                 </Toolbar>
             </AppBar>
         </Box>
