+++ src/features/auth/LoginButton.tsx
@@ -1,22 +1,14 @@
 import React from 'react';
 import Button from '@mui/material/Button';
-import { RootState } from '../../store';
-import { useDispatch, useSelector } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import axios from 'axios';
-import { resetAuth, setUserTokenAuth } from './userSlice';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
+import { setUserTokenAuth } from './userSlice';
 import { addAlert } from '../alerts/alertsSlice';
 import LoginDialog from './LoginDialog';
 
 type Props = {};
 
 const LoginButton = (props: Props) => {
-    const authed = useSelector(
-        (state: RootState) =>
-            state.authReducer.token != null && !state.authReducer.loading
-    );
-
     const [open, setOpen] = React.useState(false);
 
     const dispatch = useDispatch();
@@ -62,13 +54,9 @@ const LoginButton = (props: Props) => {
 
     return (
         <React.Fragment>
-            {authed ? (
-                <MenuButton />
-            ) : (
-                <Button color="inherit" onClick={() => setOpen(true)}>
-                    Login
-                </Button>
-            )}
+            <Button color="inherit" onClick={() => setOpen(true)}>
+                Login
+            </Button>
             <LoginDialog
                 isOpen={open}
                 setOpen={setOpen}
@@ -79,57 +67,3 @@ const LoginButton = (props: Props) => {
 };
 
 export default LoginButton;
-
-const MenuButton = () => {
-    const fullName = useSelector(
-        (state: RootState) => state.authReducer.user?.full_name
-    );
-
-    const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
-    const open = Boolean(anchorEl);
-
-    const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    };
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
-
-    const dispatch = useDispatch();
-    const logoutCallback = React.useCallback(() => {
-        dispatch(resetAuth());
-        dispatch(
-            addAlert({
-                message: 'Logged out!',
-                type: null,
-            })
-        );
-    }, [dispatch]);
-
-    return (
-        <React.Fragment>
-            <Button
-                id="basic-button"
-                aria-controls={open ? 'basic-menu' : undefined}
-                aria-haspopup="true"
-                aria-expanded={open ? 'true' : undefined}
-                onClick={handleClick}
-                color="inherit"
-            >
-                {fullName}
-            </Button>
-            <Menu
-                id="basic-menu"
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                    'aria-labelledby': 'basic-button',
-                }}
-            >
-                <MenuItem onClick={handleClose}>Profile</MenuItem>
-                <MenuItem onClick={logoutCallback}>Logout</MenuItem>
-            </Menu>
-        </React.Fragment>
-    );
-};
